Move backgroundmode state to userStore & add type
