+++ client/src/scenes/Background.ts
@@ -1,5 +1,5 @@
 import Phaser from 'phaser'
-import { BackgroundMode } from '../util'
+import { BackgroundMode } from '../../../types/BackgroundMode'
 
 export default class Background extends Phaser.Scene {
   private cloud!: Phaser.Physics.Arcade.Group
@@ -11,8 +11,8 @@ export default class Background extends Phaser.Scene {
   }
 
   create(data: { backgroundMode: BackgroundMode }) {
-    const sceneHeight = window.innerHeight
-    const sceneWidth = window.innerWidth
+    const sceneHeight = this.cameras.main.height
+    const sceneWidth = this.cameras.main.width
 
     // set texture of images based on the background mode
     if (data.backgroundMode === BackgroundMode.DAY) {
@@ -26,20 +26,12 @@ export default class Background extends Phaser.Scene {
     }
 
     // Add backdrop image
-    const backdropImage = this.add.image(
-      this.cameras.main.width / 2,
-      this.cameras.main.height / 2,
-      this.backdropKey
-    )
+    const backdropImage = this.add.image(sceneWidth / 2, sceneHeight / 2, this.backdropKey)
     const scale = Math.max(sceneWidth / backdropImage.width, sceneHeight / backdropImage.height)
     backdropImage.setScale(scale).setScrollFactor(0)
 
     // Add sun or moon image
-    const sunMoonImage = this.add.image(
-      this.cameras.main.width / 2,
-      this.cameras.main.height / 2,
-      'sun_moon'
-    )
+    const sunMoonImage = this.add.image(sceneWidth / 2, sceneHeight / 2, 'sun_moon')
     const scale2 = Math.max(sceneWidth / sunMoonImage.width, sceneHeight / sunMoonImage.height)
     sunMoonImage.setScale(scale2).setScrollFactor(0)
