+++ client/src/stores/UserStore.ts
@@ -1,15 +1,33 @@
 import { createSlice, PayloadAction } from '@reduxjs/toolkit'
 import { sanitizeId } from '../util'
+import { BackgroundMode } from '../../../types/BackgroundMode'
+
+import phaserGame from '../PhaserGame'
+import Bootstrap from '../scenes/Bootstrap'
+
+export function getInitialBackgroundMode() {
+  const currentHour = new Date().getHours()
+  return currentHour > 6 && currentHour <= 18 ? BackgroundMode.DAY : BackgroundMode.NIGHT
+}
 
 export const userSlice = createSlice({
   name: 'user',
   initialState: {
+    backgroundMode: getInitialBackgroundMode(),
     sessionId: '',
     videoConnected: false,
     loggedIn: false,
     playerNameMap: new Map<string, string>(),
   },
   reducers: {
+    switchBackgroundMode: (state) => {
+      const newMode =
+        state.backgroundMode === BackgroundMode.DAY ? BackgroundMode.NIGHT : BackgroundMode.DAY
+
+      state.backgroundMode = newMode
+      const bootstrap = phaserGame.scene.keys.bootstrap as Bootstrap
+      bootstrap.changeBackgroundMode(newMode)
+    },
     setSessionId: (state, action: PayloadAction<string>) => {
       state.sessionId = action.payload
     },
@@ -29,6 +47,7 @@ export const userSlice = createSlice({
 })
 
 export const {
+  switchBackgroundMode,
   setSessionId,
   setVideoConnected,
   setLoggedIn,
