+++ types/BackgroundMode.ts
@@ -1,0 +1,4 @@
+export enum BackgroundMode {
+  DAY,
+  NIGHT,
+}
